Rename common -> geo-cell-layer
